Add files via upload
this text file contains the link to download the fire.zip dataset for Team 25's EndSem DL project, since Github doesn't allow the zip file of such a size to be uploaded here. kindly download the fire.zip dataset from this link and save it in the running environment of the python code before running the DL Endsem Project code.